Move getSuites to common helpers file.
